hash: make gfni stubs inline
This reverts commit 07d836e5929d18ad6640ebae90dd2f81a2cafb71.

Tyler found build issues with MSVC and the thash gfni stubs.
The problem would be link errors from missing symbols.

The purpose of the original commit was to allow local definition of
RTE_LOGTYPE_HASH. Put the thash gfni back as inlines, but require
that the header only be included inside rte_thash.h so that the
log macros are defined.

Reported-by: Tyler Retzlaff <roretzla@linux.microsoft.com>
Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>