net/hns3: fix user priority 4-7 didn't enable PFC
Currently, hns3 driver only enable user priority 0-3 when the
user set TC number to 4 and enable PFC and ETS by dev_configure.
In this moment, the packet with user priority 4-7 cannot trigger
PFC frame. By the way, the nb_tcs from user can never be 0 because
of the ahead check in driver. So remove this redundant code.

Fixes: 62e3ccc2b94c ("net/hns3: support flow control")
Cc: stable@dpdk.org

Signed-off-by: Jie Hai <haijie1@huawei.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>