net/nfp: fix failure of PF initiation on flower firmware
The previous logic did not distinguish the firmware type when
reading the value of "app_cap" during the initialization of the
PF. This operation is not supported by the flower firmware, and
this will lead to a startup failure.
Fix this by adding the logic to check the firmware type.

Fixes: c7a6970fc637 ("net/nfp: notify BSP if FW speed needs to be switched")
Cc: stable@dpdk.org

Signed-off-by: Zerun Fu <zerun.fu@corigine.com>
Reviewed-by: Chaoyong He <chaoyong.he@corigine.com>
Reviewed-by: Long Wu <long.wu@corigine.com>
Reviewed-by: Peng Zhang <peng.zhang@corigine.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>