doc: update required Mingw version
Add a required version for mingw. The version is selected is
the version available for Debian oldstable (Bullseye) as listed on:
  https://www.mingw-w64.org/downloads/

The 10.0 version was released in 4/4/2022,
The current upstream version 11.0.

Also drop note about Ubuntu 16.04 because that release is EOL
and the if using a current version of Ubuntu, the package will
not be a problem.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>