eal: allow user to set default log stream before init
It is useful for application to be able to set the default log
stream before call rte_eal_init(). This makes all messages go
to the new default.

For example, to skip using syslog; just doing
	rte_openlog_stream(stderr);

There is no reason for helper command line applications to clutter
syslog with messages.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>