net/mlx5/hws: fix rule resize status check
The check to detect if a rule is in resize was done incorrectly,
this can lead to an incorrect function call upon completion for
rules which are not in resize (move).
Since the resize_info is in a union we cannot rely only on the
pointer value but also need to make sure the matcher is in resize
process, this assures us the resize info is valid.

Fixes: 762feceb8294 ("net/mlx5/hws: support resizable matchers")

Signed-off-by: Yevgeny Kliteynik <kliteyn@nvidia.com>
Signed-off-by: Alex Vesker <valex@nvidia.com>
Acked-by: Matan Azrad <matan@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>