net/tap: stop "vendoring" linux bpf headers
The proper place for finding bpf structures and functions is
in linux/bpf.h. The original version was trying to workaround the
case where the build environment was running on old pre BPF
version of Glibc, but the target environment had BPF.

Having own private (and divergent) version headers leads to future
problems when BPF definitions evolve.

This also blocks tap device from building on kernels before 4.2.
The oldest supported LTS kernel is currently 4.4 and the 4.1
kernel reached end of life in December 2015.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>