net/axgbe: reset link when the link never comes back
Normally, auto negotiation and reconnect should be automatically done by
the hardware. But there seems to be an issue where auto negotiation has
to be restarted manually. This happens because of link training and so
even though still connected to the partner the link never "comes back".
This needs an auto-negotiation restart.

Signed-off-by: Venkat Kumar Ande <VenkatKumar.Ande@amd.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>