net/axgbe: improve the SFP codes check for DAC cables
The current AXGBE code assumes that offset 6 of EEPROM SFP DAC (passive)
cables is NULL. However, some cables (the 5 meter and 7 meter Molex
passive cables) have non-zero data at offset 6. Fix the logic by moving
the passive cable check above the active checks, so as not to be
improperly identified as an active cable. This will fix the issue for
any passive cable that advertises 1000Base-CX in offset 6.

Signed-off-by: Venkat Kumar Ande <VenkatKumar.Ande@amd.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>