net/axgbe: improve logic around active and passive cables
SFP+ active and passive cables are copper cables with fixed SFP+ end
connectors. Due to a misinterpretation of this, SFP+ active cables could
end up not being recognized, causing the driver to fail to establish a
connection.

Introduce a new enum in SFP+ cable types, XGBE_SFP_CABLE_FIBER, that is
the default cable type, and handle active and passive cables when they are
specifically detected.

Signed-off-by: Venkat Kumar Ande <VenkatKumar.Ande@amd.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>