net/axgbe: improve the false linkup in axgbe PHY status
In the event of a change in AXGBE mode, the current auto-negotiation
needs to be reset and the AN cycle needs to be re-triggerred. However,
the current code ignores the return value of axgbe_set_mode(), leading to
false information as the link is declared without checking the status
register.

Fix this by propagating the mode switch status information to
axgbe_phy_status().

Signed-off-by: Venkat Kumar Ande <VenkatKumar.Ande@amd.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>