fix semicolon at the end of RTE_LOG_REGISTER_DEFAULT
The macro RTE_LOG_REGISTER_DEFAULT emits code for an initialization
function. If a driver (and most do) adds a semicolon after the macro.

RTE_LOG_REGISTER_DEFAULT(logtype_foo, INFO);

Is equivalent to:

	int logtype_foo;
	static void __logtype_foo(void) {
	    logtype_foo = rte_log_register_type_and_pick_level(name,
	                  	 RTE_LOG_INFO);
		if (type < 0)
			logtype_foo = RTE_LOGTYPE_EAL;
	};
The problem is that extra semi-colon after the function.

If code is built with highest level of warnings (pedantic),
then it will generate a warning.
  warning: ISO C does not allow extra ‘;’ outside of a function

This is a treewide fix for this.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Signed-off-by: 0-day Robot <robot@bytheb.org>